backend/requirements.txt
fastapi
uvicorn
lark-oapi
python-dotenv
sqlalchemy
APScheduler
